Moved: restructure modal files and update import paths for cart and login modals
