+++ composables/useConnectService.ts
@@ -1,4 +1,4 @@
-import {ConnectService} from "~/modals/connect.modal";
+import {ConnectService} from "~/domains/connect/connect.modal";
 
 export const useConnectService = () => {
   return new ConnectService('http://localhost:8080', useToastService())
